requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain-text-splitters
faiss-cpu
sentence-transformers
python-dotenv
tqdm
